Moved src files Added auto download and extraction of model file

+++ download_models.ps1
@@ -1,18 +1,18 @@
-function Expand-Tar($tarFile, $dest) {
-
-    if (-not (Get-Command Expand-7Zip -ErrorAction Ignore)) {
-        Install-Package -Scope CurrentUser -Force 7Zip4PowerShell > $null
-    }
-
-    Expand-7Zip $tarFile $dest
+$ErrorActionPreference = "Stop"
+Import-Module BitsTransfer
+if(!(Test-Path -Path "bin\server\models\ssd_inception_v2_coco_2017_11_17\")){
+  new-item -Name bin\server\models -Force -ItemType directory
+  Start-BitsTransfer -Source "http://download.tensorflow.org/models/object_detection/ssd_inception_v2_coco_2017_11_17.tar.gz" -Destination "ssd_inception_v2_coco_2017_11_17.tar.gz"
+  .\tools\7za.exe e "ssd_inception_v2_coco_2017_11_17.tar.gz"
+  .\tools\7za.exe x "ssd_inception_v2_coco_2017_11_17.tar" -obin\server\models\ -r
+  Remove-Item -Path ssd_inception_v2_coco_2017_11_17.tar
+  Remove-Item -Path ssd_inception_v2_coco_2017_11_17.tar.gz
 }
-
-new-item -Name bin\server\models -ItemType directory
-
-$client = new-object System.Net.WebClient
-$client.DownloadFile("http://download.tensorflow.org/models/object_detection/ssd_inception_v2_coco_2017_11_17.tar.gz","ssd_inception_v2_coco_2017_11_17.tar.gz")
-Expand-Tar ssd_inception_v2_coco_2017_11_17.tar.gz bin\server\models
-Remove-Item -Path ssd_inception_v2_coco_2017_11_17.tar.gz
-new-item -Name bin\server\labels -ItemType directory
-$client.DownloadFile("https://raw.githubusercontent.com/tensorflow/models/865c14c1209cb9ae188b2a1b5f0883c72e050d4c/research/object_detection/data/mscoco_label_map.pbtxt", "bin\server\labels\mscoco_label_map.pbtxt")
-$client.DownloadFile("https://raw.githubusercontent.com/tensorflow/models/865c14c1209cb9ae188b2a1b5f0883c72e050d4c/research/object_detection/data/oid_bbox_trainable_label_map.pbtxt", "bin\server\labels\oid_bbox_trainable_label_map.pbtxt") 
+if(!(Test-Path "bin\server\labels\mscoco_label_map.pbtxt" -PathType Leaf)){
+  new-item -Name bin\server\labels -Force -ItemType directory
+  Start-BitsTransfer -Source "https://raw.githubusercontent.com/tensorflow/models/865c14c1209cb9ae188b2a1b5f0883c72e050d4c/research/object_detection/data/mscoco_label_map.pbtxt" -Destination  "bin\server\labels\mscoco_label_map.pbtxt"
+}
+if(!(Test-Path "bin\server\labels\oid_bbox_trainable_label_map.pbtxt" -PathType Leaf)){
+  new-item -Name bin\server\labels -Force -ItemType directory
+  Start-BitsTransfer -Source "https://raw.githubusercontent.com/tensorflow/models/865c14c1209cb9ae188b2a1b5f0883c72e050d4c/research/object_detection/data/oid_bbox_trainable_label_map.pbtxt" -Destination  "bin\server\labels\oid_bbox_trainable_label_map.pbtxt"
+}
